+++ src/main/kotlin/com/github/asoee/cursorlessjetbrains/cursorless/CursorlessCallback.kt
@@ -14,4 +14,5 @@ interface CursorlessCallback {
     fun executeRangeCommand(editorId: String, rangeCommand: CursorlessEditorCommand)
     fun revealLine(editorId: String, line: Int, revealAt: String)
     fun flashRanges(flashRanges: Array<CursorlessFlashRange>)
+    fun prePhraseVersion(): String?
 }