+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 
 ## [Unreleased]
 
+## [0.0.6] - 2024-12-18
+
 ### Fixed
 
 - Fix indentation and cursor position issues after cursorless `pour` command
@@ -69,7 +71,8 @@
 - Communication with talon using both HTTP and file based command server
 - Most voice-code commands added to the plugin
 
-[Unreleased]: https://github.com/asoee/cursorless-jetbrains/compare/v0.0.5...HEAD
+[Unreleased]: https://github.com/asoee/cursorless-jetbrains/compare/v0.0.6...HEAD
+[0.0.6]: https://github.com/asoee/cursorless-jetbrains/compare/v0.0.5...v0.0.6
 [0.0.5]: https://github.com/asoee/cursorless-jetbrains/compare/v0.0.4...v0.0.5
 [0.0.4]: https://github.com/asoee/cursorless-jetbrains/compare/v0.0.3...v0.0.4
 [0.0.3]: https://github.com/asoee/cursorless-jetbrains/compare/v0.0.2...v0.0.3